Refector: 전체 감가액을 Float -> Double

+++ src/main/kotlin/com/hwalaon/wezxro_server/domain/account/controller/response/AccountDetailResponse.kt
@@ -3,13 +3,13 @@ package com.hwalaon.wezxro_server.domain.account.controller.response
 import com.hwalaon.wezxro_server.domain.account.model.Account
 import com.hwalaon.wezxro_server.global.common.basic.constant.BasicStatus
 
-data class AccountDetailResponse (
+data class AccountDetailResponse(
     val userId: Long,
     val name: String,
     val email: String,
     val money: Double,
     val status: BasicStatus,
-    val staticRate: Float,
+    val staticRate: Double,
 ) {
     companion object {
         fun fromDomain(domain: Account) =
